+++ src/briefy/choreographer/__init__.py
@@ -20,5 +20,19 @@ if LOG_SERVER:
         LOG_SERVER, 5543, version=1, tags=['Worker', 'briefy.choreographer']
     )
     logger.addHandler(log_handler)
+    from briefy.common.utils.transformers import json_dumps
+    from logstash import formatter
+
+    # custom json wrapper do serialize complex objects and monkey patch logstash.formater.json
+    class CustomJSON:
+        """Custom json wrapper."""
+
+        @staticmethod
+        def dumps(message):
+            """Dump a message to json."""
+            return json_dumps(message)
+
+    formatter.json = CustomJSON
+
 
 XMLConfig('configure.zcml', choreographer)()